+++ src/vscode/editors/DatabricksSendToAPI.ts
@@ -1,0 +1,12 @@
+import * as vscode from 'vscode';
+
+export class DatabricksSendToAPI {
+	
+	static async sendToAPI(resourceUri: vscode.Uri): Promise<void> {
+		// TODO dropdown(s) for different APIs / endpoints
+		// TODO dropdown for different methods (GET/POST/PUT/DELETE/...)
+		// TODO send editor content as workload to API
+		let x = 0;
+	}
+}
+