+++ src/vscode/notebook/Widgets/DatabricksTextWidget.ts
@@ -4,19 +4,35 @@ import { DatabricksWidget } from './DatabricksWidget';
 
 
 export class DatabricksTextWidget extends DatabricksWidget<string> {
-	private static WidgetRegEx = /dbutils\.widgets\.text\s*\(["']{1}(?<name>.*?)["']{1}\s*,\s*["']{1}(?<defaultValue>.*?)["'](,\s*["']{1}(?<label>.*?)["'])?\)/gm;
 
+	private static WidgetRegExPositional: RegExp = /dbutils\.widgets\.text\(["']{1}(?<name>.*?)["']{1}\s*,\s*["']{1}(?<default>.*?)["'](,\s*["']{1}(?<label>.*?)["'])?\)/gm;
+	private static WidgetRegExNamed: RegExp = /dbutils\.widgets\.text\((.*\s*name\s*=\s*['"](?<name>.*?)['"]\s*)(.*\s*defaultValue\s*=\s*['"](?<default>.*?)['"]\s*)(.*\s*label\s*=\s*['"](?<label>.*?)['"]\s*)\)/gm;
 
 	constructor(language: ContextLanguage, name: string, defaultValue?: string, label?: string) {
 		super(language, "text", name, defaultValue, label);
 	}
 
 	static loadFromCommandText(commandText: string, language: ContextLanguage): DatabricksTextWidget[] {
-		let matches = commandText.matchAll(DatabricksTextWidget.WidgetRegEx);
+		let widgets: DatabricksTextWidget[] = [];
+
+		widgets = widgets.concat(this.parseRegEx(commandText, this.WidgetRegExPositional, language));
+		widgets = widgets.concat(this.parseRegEx(commandText, this.WidgetRegExNamed, language));
+
+
+		return widgets;
+	}
+
+	private static parseRegEx(commandText: string, regex: RegExp, language: ContextLanguage): DatabricksTextWidget[] {
+		const matches = commandText.matchAll(regex);
 
 		let widgets: DatabricksTextWidget[] = [];
+
 		for (let match of matches) {
-			let widget = new DatabricksTextWidget(language, match.groups["name"], match.groups["defaultValue"], match.groups["label"]);
+			let widget = new DatabricksTextWidget(language,
+				match.groups["name"],
+				match.groups["default"],
+				match.groups["label"]
+			);
 			widgets.push(widget);
 		}