+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - introduced a new notebook type for `Scala`, `SQL` and `R`
 - removed interactive notebook kernel as it never really worked
   - replaced it with new Databricks kernel to work with the new notebook type
+- fix issue with named parameters in widgets ([#128](/../../issues/128))
 - minor fix to handle `pip install` more generic
 
 **v2.0.3:**